Fixed restaurant mockup regeneration problem
Found an issue where an object was referencing another object
rather than creating its own copy.  I fixed the problem by converting
the object to JSON and then back to a string.  While this isn't the best
way to copy an object, it does a better job at simulating what I would
have to do with an actual AJAX call.